Fixed some bugs
FIXED-When editing an employee, it is possible to set administrator status to things besides 1 or 0.
FIXED-When logged in as moderator, not all buttons work.
FOUND GLITCH (TEMP FIX)- Inventory button doesn't work. Temp fixed by linking it to 'EDIT ITEM' button. Recommend either Edit Item button is removed, or functionality is fixed. As of fix, both buttons will do the same thing.